moved title tests to test_figure
